requirements.txt
asttokens==2.0.5
backcall==0.2.0
boto3==1.23.6
botocore==1.26.6
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jmespath==1.0.0
jupyter-client==7.3.1
jupyter-core==4.10.0
Markdown==3.3.7
matplotlib-inline==0.1.3
mysql-connector==2.2.9
mysqlclient==2.1.0
nest-asyncio==1.5.5
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
Pygments==2.12.0
pyodbc==4.0.32
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pyzmq==23.0.0
s3transfer==0.5.2
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.2.1.post0
urllib3==1.26.9
wcwidth==0.2.5